+++ app/js/homeydash.app.js
@@ -1,4 +1,4 @@
-var version = "1.1.2"
+var version = "1.1.3"
 
 var CLIENT_ID = '5cbb504da1fc782009f52e46';
 var CLIENT_SECRET = 'gvhs0gebgir8vz8yo2l0jfb49u9xzzhrkuo1uvs8';
@@ -233,6 +233,7 @@ window.addEventListener('load', function() {
   if ( logofromurl == undefined ) { logofromurl = "" }
 
   var zoom = getCookie("zoom")
+
   $content.style.zoom = zoom;
 
   var token = getQueryVariable('token');
@@ -383,6 +384,16 @@ window.addEventListener('load', function() {
               }
             });
           }
+          if ( device.capabilitiesObj.locked ) {
+            device.makeCapabilityInstance('locked', function(value){
+              var $valueElement = document.getElementById('lock:' + device.id);
+              if( $valueElement ) {
+                console.log("Locked: " + value)
+                $valueElement.classList.toggle('locked', !!value);
+                $valueElement.classList.toggle('unlocked', !value);
+              }
+            });
+          }
           if ( device.capabilitiesObj.alarm_generic ) {
             device.makeCapabilityInstance('alarm_generic', function(value){
               var $deviceElement = document.getElementById('device:' + device.id);
@@ -504,6 +515,16 @@ window.addEventListener('load', function() {
               if( $deviceElement ) {
                 var $valueElement = document.getElementById('value:' + device.id + ":measure_rain");
                 capability = device.capabilitiesObj['measure_rain']
+                renderValue($valueElement, capability.id, capability.value, capability.units)
+              }
+            });
+          }
+          if ( device.capabilitiesObj.measure_rain_day ) {
+            device.makeCapabilityInstance('measure_rain_day', function(value){
+              var $deviceElement = document.getElementById('device:' + device.id);
+              if( $deviceElement ) {
+                var $valueElement = document.getElementById('value:' + device.id + ":measure_rain_day");
+                capability = device.capabilitiesObj['measure_rain_day']
                 renderValue($valueElement, capability.id, capability.value, capability.units)
               }
             });
@@ -744,17 +765,9 @@ window.addEventListener('load', function() {
       $versionIcon.addEventListener('click', function() {
         setCookie('version', version ,12)
         changeLog = ""
-        changeLog = changeLog + "* Added option to change order of devices/flows/alarms<br />"
-        changeLog = changeLog + "* Added option to dim on supported devices<br />"
-        changeLog = changeLog + "* minor css fixes<br />"
-        changeLog = changeLog + "* fixed setting wrong brightness on closing slider<br />"
-        changeLog = changeLog + "* fixed rounding error in temperature<br />"
-        changeLog = changeLog + "* Added option to change target temperature on thermostat devices<br />"
-        changeLog = changeLog + "* Added option to change volume on audio devices<br />"
-        changeLog = changeLog + "* Tweaked placement of slider<br />"
-        changeLog = changeLog + "* Added Italian translations<br />"
-        changeLog = changeLog + "* Added WeatherFlow's Smart Weather Station capabilities<br />"
-        
+        changeLog = changeLog + "* Added lock status icon<br />"
+        changeLog = changeLog + "* Prevent unloadable device icons causing display problems<br />"
+        changeLog = changeLog + "* Corrected Italian translations<br />"
         renderInfoPanel("u",changeLog)
       })
     }
@@ -1137,7 +1150,11 @@ window.addEventListener('load', function() {
       var $icon = document.createElement('div');
       $icon.id = 'icon:' + device.id
       $icon.classList.add('icon');
-      $icon.style.webkitMaskImage = 'url(https://icons-cdn.athom.com/' + device.iconObj.id + '-128.png)';
+      if ( device.iconObj ) {
+        $icon.style.webkitMaskImage = 'url(https://icons-cdn.athom.com/' + device.iconObj.id + '-128.png)';
+      } else if ( device.icon ) {
+        $icon.style.webkitMaskImage ='url(img/capabilities/blank.png)';
+      }
       if ( device.name == "Bier" || device.name == "Bier temperatuur" ) {
         $icon.style.webkitMaskImage = 'url(img/capabilities/beer.png)';
         $icon.style.backgroundImage = 'url(img/capabilities/beer.png)';
@@ -1165,6 +1182,19 @@ window.addEventListener('load', function() {
             renderValue($value, capability.id, capability.value, capability.units)
             if (device.name=="Bier") {renderValue($value, capability.id, capability.value, "")}
             $deviceElement.appendChild($value)
+            itemNr =itemNr + 1
+          } else 
+          if ( capability.id == "locked" ) {
+            var $lock = document.createElement('div');
+            $lock.id = 'lock:' + device.id
+            $lock.title = capability.title
+            $lock.classList.add('icon-capability-lock');
+            if ( device.capabilitiesObj.locked.value ) {
+              $lock.classList.add('locked');
+            } else {
+              $lock.classList.add('unlocked');
+            }
+            $deviceElement.appendChild($lock)
             itemNr =itemNr + 1
           }
         }
@@ -1299,7 +1329,8 @@ window.addEventListener('load', function() {
         capabilityId == "measure_humidity"
         ) {
       capabilityValue = Math.round(capabilityValue*10)/10
-      var integer = Math.floor(capabilityValue)
+      //var integer = Math.floor(capabilityValue)
+      var integer = parseInt(capabilityValue)
       n = Math.abs(capabilityValue)
       var decimal = Math.round((n - Math.floor(n))*10)/10 + "-"
       var decimal = decimal.substring(2,3)
@@ -1585,8 +1616,9 @@ window.addEventListener('load', function() {
           capabilityToShow = capability.id
           // measure_uv and measure_solarradiation icons are broken at icons-cdn.athom.com
           if ( capability.iconObj && capability.id != "measure_uv" && capability.id != "measure_solarradiation" ) {
+          //if ( capability.iconObj ) {
             iconToShow = 'https://icons-cdn.athom.com/' + capability.iconObj.id + '-128.png'
-
+            console.log(iconToShow)
           } else {
             iconToShow = 'img/capabilities/' + capability.id + '.png'
           }